+++ src/addons/display_commander/exit_handler.cpp
@@ -5,7 +5,6 @@
 #include <reshade.hpp>
 #include <sstream>
 #include <vector>
-#include "features/presentmon/presentmon_minimal_etw.hpp"
 #include "settings/main_tab_settings.hpp"
 #include "utils/general_utils.hpp"
 #include "hooks/loadlibrary_hooks.hpp"
@@ -70,7 +69,6 @@ void OnHandleExit(ExitSource source, const std::string& message) {
 
     //display_commander::config::DisplayCommanderConfigManager::GetInstance().SetAutoFlushLogs(true);
     display_commander::logger::FlushLogs();
-    display_commander::features::presentmon::ShutdownPresentMonEtw();
 
 #if defined(DC_EXTERNAL_MODULES)
     modules::presentmon_module::ShutdownPresentMonForProcessExit();